captioning-data-engine/requirements.txt
accelerate==0.33.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
asyncer==0.0.2
attrs==24.2.0
bidict==0.23.1
bitsandbytes==0.43.2
certifi==2024.7.4
chainlit==1.1.400
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.20.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.13.1
filetype==1.2.0
fire==0.6.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.3
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
hyperframe==6.0.1
idna==3.7
importlib_metadata==8.0.0
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
Lazify==0.4.0
literalai==0.0.607
lmdeploy==0.5.3
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
mdurl==0.1.2
mmengine-lite==0.10.4
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
openai==1.37.1
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
packaging==23.2
pandas==2.2.2
peft==0.11.1
pillow==10.2.0
platformdirs==4.2.2
portalocker==2.10.1
protobuf==4.25.4
psutil==6.0.0
pyarrow==17.0.0
pyarrow-hotfix==0.6
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
qdrant-client==1.11.0
regex==2024.7.24
requests==2.32.3
rich==13.7.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sentence-transformers==3.0.1
sentencepiece==0.2.0
shortuuid==1.0.13
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sse-starlette==2.1.2
starlette==0.37.2
sympy==1.12
syncer==2.0.3
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.8
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchaudio==2.4.0+cu124
torchvision==0.19.0
tqdm==4.66.5
transformers==4.40.0
triton==2.3.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uptrace==1.26.0
urllib3==2.2.2
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.27.post2
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.20.0
